Refactor mergecheck
